+++ CHANGELOG.md
@@ -21,9 +21,13 @@ FlatLaf Change Log
 - TabbedPane: Support closable tabs. (PR #193; issues #31 and #40)
 - TabbedPane: Support minimum or maximum tab widths. (set client property
   `JTabbedPane.minimumTabWidth` or `JTabbedPane.maximumTabWidth` to an integer)
+  (PR #199)
 - TabbedPane: Support alignment of tab area. (set client property
   `JTabbedPane.tabAreaAlignment` to `"leading"`, `"trailing"`, `"center"` or
-  `"fill"`)
+  `"fill"`) (PR #199)
+- TabbedPane: Support equal and compact tab width modes. (set client property
+  `JTabbedPane.tabWidthMode` to `"preferred"`, `"equal"` or `"compact"`) (PR
+  #199)
 - Support painting separator line between window title and content (use UI value
   `TitlePane.borderColor`). (issue #184)
 - Extras: `FlatSVGIcon` now allows specifying icon width and height in